+++ simple_history/manager.py
@@ -1,10 +1,11 @@
-from __future__ import unicode_literals
+from django.db import connection, models
+from django.db.models import OuterRef, Subquery
+from django.utils import timezone
 
-from django.db import models
-from django.utils.timezone import now
+from simple_history.utils import get_change_reason_from_object
 
 
-class HistoryDescriptor(object):
+class HistoryDescriptor:
     def __init__(self, model):
         self.model = model
 
@@ -89,30 +90,78 @@ class HistoryManager(models.Manager):
         model = type(self.model().instance)  # a bit of a hack to get the model
         pk_attr = model._meta.pk.name
         queryset = self.get_queryset().filter(history_date__lte=date)
-        for original_pk in set(queryset.order_by().values_list(pk_attr, flat=True)):
-            changes = queryset.filter(**{pk_attr: original_pk})
-            last_change = changes.latest("history_date")
-            if changes.filter(
-                history_date=last_change.history_date, history_type="-"
-            ).exists():
-                continue
-            yield last_change.instance
+        # If using MySQL, need to get a list of IDs in memory and then use them for the
+        # second query.
+        # Does mean two loops through the DB to get the full set, but still a speed
+        # improvement.
+        backend = connection.vendor
+        if backend == "mysql":
+            history_ids = {}
+            for item in queryset.order_by("-history_date", "-pk"):
+                if getattr(item, pk_attr) not in history_ids:
+                    history_ids[getattr(item, pk_attr)] = item.pk
+            latest_historics = queryset.filter(history_id__in=history_ids.values())
+        elif backend == "postgresql":
+            latest_pk_attr_historic_ids = (
+                queryset.order_by(pk_attr, "-history_date", "-pk")
+                .distinct(pk_attr)
+                .values_list("pk", flat=True)
+            )
+            latest_historics = queryset.filter(
+                history_id__in=latest_pk_attr_historic_ids
+            )
+        else:
+            latest_pk_attr_historic_ids = (
+                queryset.filter(**{pk_attr: OuterRef(pk_attr)})
+                .order_by("-history_date", "-pk")
+                .values("pk")[:1]
+            )
+            latest_historics = queryset.filter(
+                history_id__in=Subquery(latest_pk_attr_historic_ids)
+            )
+        adjusted = latest_historics.exclude(history_type="-").order_by(pk_attr)
+        for historic_item in adjusted:
+            yield historic_item.instance
 
-    def bulk_history_create(self, objs, batch_size=None):
-        """Bulk create the history for the objects specified by objs"""
+    def bulk_history_create(
+        self,
+        objs,
+        batch_size=None,
+        update=False,
+        default_user=None,
+        default_change_reason="",
+        default_date=None,
+    ):
+        """
+        Bulk create the history for the objects specified by objs.
+        If called by bulk_update_with_history, use the update boolean and
+        save the history_type accordingly.
+        """
+
+        history_type = "+"
+        if update:
+            history_type = "~"
 
         historical_instances = []
         for instance in objs:
+            history_user = getattr(
+                instance,
+                "_history_user",
+                default_user or self.model.get_default_history_user(instance),
+            )
             row = self.model(
-                history_date=getattr(instance, "_history_date", now()),
-                history_user=getattr(instance, "_history_user", None),
-                history_change_reason=getattr(instance, "changeReason", ""),
-                history_type="+",
+                history_date=getattr(
+                    instance, "_history_date", default_date or timezone.now()
+                ),
+                history_user=history_user,
+                history_change_reason=get_change_reason_from_object(instance)
+                or default_change_reason,
+                history_type=history_type,
                 **{
                     field.attname: getattr(instance, field.attname)
                     for field in instance._meta.fields
                     if field.name not in self.model._history_excluded_fields
-                }
+                },
             )
             if hasattr(self.model, "history_relation"):
                 row.history_relation_id = instance.pk