+++ simple_history/signals.py
@@ -1,23 +1,9 @@
 import django.dispatch
 
+# Arguments: "instance",  "history_instance", "history_date",
+#            "history_user", "history_change_reason", "using"
+pre_create_historical_record = django.dispatch.Signal()
 
-pre_create_historical_record = django.dispatch.Signal(
-    providing_args=[
-        "instance",
-        "history_instance",
-        "history_date",
-        "history_user",
-        "history_change_reason",
-        "using",
-    ]
-)
-post_create_historical_record = django.dispatch.Signal(
-    providing_args=[
-        "instance",
-        "history_instance",
-        "history_date",
-        "history_user",
-        "history_change_reason",
-        "using",
-    ]
-)
+# Arguments: "instance",  "history_instance", "history_date",
+#            "history_user", "history_change_reason", "using"
+post_create_historical_record = django.dispatch.Signal()