+++ simple_history/tests/admin.py
@@ -1,9 +1,8 @@
-from __future__ import unicode_literals
-
 from django.contrib import admin
 
 from simple_history.admin import SimpleHistoryAdmin
 from simple_history.tests.external.models import ExternalModelWithCustomUserIdField
+
 from .models import (
     Book,
     Choice,
@@ -13,6 +12,7 @@ from .models import (
     FileModel,
     Paper,
     Person,
+    Planet,
     Poll,
 )
 
@@ -33,6 +33,13 @@ class FileModelAdmin(SimpleHistoryAdmin):
     history_list_display = ["title", "test_method"]
 
 
+class PlanetAdmin(SimpleHistoryAdmin):
+    def test_method(self, obj):
+        return "test_method_value"
+
+    history_list_display = ["title", "test_method"]
+
+
 admin.site.register(Poll, SimpleHistoryAdmin)
 admin.site.register(Choice, ChoiceAdmin)
 admin.site.register(Person, PersonAdmin)
@@ -43,3 +50,4 @@ admin.site.register(Employee, SimpleHistoryAdmin)
 admin.site.register(ConcreteExternal, SimpleHistoryAdmin)
 admin.site.register(ExternalModelWithCustomUserIdField, SimpleHistoryAdmin)
 admin.site.register(FileModel, FileModelAdmin)
+admin.site.register(Planet, PlanetAdmin)