+++ simple_history/tests/models.py
@@ -1,7 +1,6 @@
-from __future__ import unicode_literals
-
 import datetime
 import uuid
+
 from django.apps import apps
 from django.conf import settings
 from django.db import models
@@ -9,11 +8,9 @@ from django.urls import reverse
 
 from simple_history import register
 from simple_history.models import HistoricalRecords
-from .custom_user.models import CustomUser as User
 
-from .external.models import AbstractExternal
-from .external.models import AbstractExternal2
-from .external.models import AbstractExternal3
+from .custom_user.models import CustomUser as User
+from .external.models import AbstractExternal, AbstractExternal2, AbstractExternal3
 
 get_model = apps.get_model
 
@@ -26,6 +23,13 @@ class Poll(models.Model):
 
     def get_absolute_url(self):
         return reverse("poll-detail", kwargs={"pk": self.pk})
+
+
+class PollWithUniqueQuestion(models.Model):
+    question = models.CharField(max_length=200, unique=True)
+    pub_date = models.DateTimeField("date published")
+
+    history = HistoricalRecords()
 
 
 class PollWithExcludeFields(models.Model):
@@ -61,6 +65,21 @@ class PollWithExcludedFKField(models.Model):
     place = models.ForeignKey("Place", on_delete=models.CASCADE)
 
     history = HistoricalRecords(excluded_fields=["place"])
+
+
+class AlternativePollManager(models.Manager):
+    def get_queryset(self):
+        return super(AlternativePollManager, self).get_queryset().exclude(id=1)
+
+
+class PollWithAlternativeManager(models.Model):
+    some_objects = AlternativePollManager()
+    all_objects = models.Manager()
+
+    question = models.CharField(max_length=200)
+    pub_date = models.DateTimeField("date published")
+
+    history = HistoricalRecords()
 
 
 class IPAddressHistoricalModel(models.Model):
@@ -192,6 +211,20 @@ class FileModel(models.Model):
     history = HistoricalRecords()
 
 
+# Set SIMPLE_HISTORY_FILEFIELD_TO_CHARFIELD
+setattr(settings, "SIMPLE_HISTORY_FILEFIELD_TO_CHARFIELD", True)
+
+
+class CharFieldFileModel(models.Model):
+    title = models.CharField(max_length=100)
+    file = models.FileField(upload_to="files")
+    history = HistoricalRecords()
+
+
+# Clear SIMPLE_HISTORY_FILEFIELD_TO_CHARFIELD
+delattr(settings, "SIMPLE_HISTORY_FILEFIELD_TO_CHARFIELD")
+
+
 class Document(models.Model):
     changed_by = models.ForeignKey(
         User, on_delete=models.CASCADE, null=True, blank=True
@@ -200,7 +233,10 @@ class Document(models.Model):
 
     @property
     def _history_user(self):
-        return self.changed_by
+        try:
+            return self.changed_by
+        except User.DoesNotExist:
+            return None
 
 
 class Paper(Document):
@@ -365,6 +401,17 @@ class City(models.Model):
         Country, on_delete=models.CASCADE, db_column="countryCode"
     )
     history = HistoricalRecords()
+
+
+class Planet(models.Model):
+    star = models.CharField(max_length=30)
+    history = HistoricalRecords()
+
+    def __str__(self):
+        return self.star
+
+    class Meta:
+        verbose_name = "Planet"
 
 
 class Contact(models.Model):
@@ -562,7 +609,6 @@ class UUIDRegisterModel(models.Model):
 
 register(UUIDRegisterModel, history_id_field=models.UUIDField(default=uuid.uuid4))
 
-
 # Set the SIMPLE_HISTORY_HISTORY_ID_USE_UUID
 setattr(settings, "SIMPLE_HISTORY_HISTORY_ID_USE_UUID", True)
 
@@ -574,7 +620,6 @@ class UUIDDefaultModel(models.Model):
 
 # Clear the SIMPLE_HISTORY_HISTORY_ID_USE_UUID
 delattr(settings, "SIMPLE_HISTORY_HISTORY_ID_USE_UUID")
-
 
 # Set the SIMPLE_HISTORY_HISTORY_CHANGE_REASON_FIELD
 setattr(settings, "SIMPLE_HISTORY_HISTORY_CHANGE_REASON_USE_TEXT_FIELD", True)
@@ -667,3 +712,19 @@ class ForeignKeyToSelfModel(models.Model):
 class Street(models.Model):
     name = models.CharField(max_length=150)
     log = HistoricalRecords(related_name="history")
+
+
+class ManyToManyModelOther(models.Model):
+    name = models.CharField(max_length=15, unique=True)
+
+
+class BulkCreateManyToManyModel(models.Model):
+    name = models.CharField(max_length=15, unique=True)
+    other = models.ManyToManyField(ManyToManyModelOther)
+    history = HistoricalRecords()
+
+
+class ModelWithExcludedManyToMany(models.Model):
+    name = models.CharField(max_length=15, unique=True)
+    other = models.ManyToManyField(ManyToManyModelOther)
+    history = HistoricalRecords(excluded_fields=["other"])