+++ simple_history/tests/tests/test_signals.py
@@ -1,12 +1,10 @@
-from __future__ import unicode_literals
-
 from datetime import datetime
 
 from django.test import TestCase
 
 from simple_history.signals import (
-    pre_create_historical_record,
     post_create_historical_record,
+    pre_create_historical_record,
 )
 
 from ..models import Poll